requirements.txt
django
psycopg2
djangorestframework
djangorestframework-simplejwt
python-dotenv
requests
redis 
django-redis
